Snapshot of copy work from today.
I expanded the precomp (under the assumption that I wanted the test to run over
a single chapel file differentiated by execution options.  Even when that stops
being true, the work here will still be valid, just split into multiple files)
to include more of the execution phases.  This meant using more unique names
for files and directory locations, but that's okay, it keeps me honest and
ensures that the execution of one does not affect the next.  I reflected the
more unique names in the execopts file.  At the moment, there are 8 more test
cases to define in the precomp and I still need to create the .goods.  And, you
know, write the function being used by this test.