Remove pragmas from chpl__autoCopy definitions.
There are two problems with this commit right now:
1) The pragma for _tuples can't be removed without causing an out of bounds
exception.  I'm not sure why that is, but it sounds like a tuple-specific
problem.  It means that the pragma cannot be turned into just a flag, which is
unfortunate
2) There is a failure with arrays/vass/const-checking.aliases.chpl where this
change for some reason prevents the constness checking on aliases from
detecting when a var alias is made of a const instance.  I don't see why this
dependency should exist, but there it is.  This problem means that this patch
cannot be merged yet.  I don't think it's worth the effort of investigating at
the moment and so will put this off until after the release.