Start of optimization of moveDir suggested by Kyle
Kyle mentioned when I committed the implementation of moveDir that it would be
more light-weight to only do a full copy+remove when crossing file system
boundaries and that otherwise a simple rename would be sufficient.  However,
the check to determine if we are operating on different filesystems shouldn't
occur on the new directory itself (since a stat on a non-existent directory
results in ENOENT rather than a valid stat struct) but rather on its parent
directory.  That path will most easily be obtained using a call to Path.dirname,
which hasn't been implemented yet.  When that get implemented, I will come
back to this change.