Add four tests of our point of instantiation rules
test/types/type_variables/pointOfInstantiation/lydia/
- methodInUsingModuleAccess.chpl

This test demonstrates the behavior when a class defined in a used module (B)
is given an additional method in the using module (A), so that if a function in
the used module (B) tries to call that method, and this function is accessed via
the using module (A), the function will compile.

- paramFuncInUsedModule.chpl

This test demonstrates the behavior when a param function defined in a used
module (B) is instantiated in the using module (A), and then a later call
which steps within the used module (B) calls the function in the form that was
already instantiated, and in a form that wasn't previously, and then the
using module (A) calls the most recently instantiated form.  The param function
makes a call to another function and, depending on where it is instantiated,
finds either A's or B's version.

- paramCalledInSiblingModule.chpl

This test demonstrates similar behavior to paramFuncInUsedModule.chpl, except
the two instantiated forms are also called by a third module (C), used by the
using module (A).

- test/visibility/private/functions/test_point_of_instantiation.chpl

This test is a variation on paramFuncInUsedModule.chpl, where the function in
the using module (A) which would otherwise be found is declared private.  In
this case, the instantiated function never finds A's version.