Start of fix to super.foo calls (timed out due to goals, shootouts)
Was going to replace the . call in normalize with the SUPER_CALL, but realized
that the AST transformations between that point in normalize and function
resolution when we insert the virtual method calls would have forced the
CallExpr to have a value in both the primitive and baseExpr field, which at the
very least our copy operation doesn't expect.  So I added the FLAG_SUPER_CLASS
to the temporary variable which will store it during normalize (since after
normalize the variable definition and the use of it get separated), and then
checked if the temp being used had that flag during function resolution when
inserting the virtual method calls.

This change will make it to codegen, but will fail then because I haven't
inserted the codegen machinery yet.