Add table of class id to parent class id into our generated code
This is implemented as an array, where the index corresponds to the id of the
class whose parent you want to access, and the element stored at that index is
the id of the parent class (or -1 if the class has no parent or has multiple
parents).  I also included a comment of the class to its parent for debugging
purposes (but only when the parent exists).

This outputs a correct table for hello.chpl.  The bug I'm fixing still fails to
compile, but that's because we don't respond to PRIM_SUPER_CALL yet.  I should
also add a check to ensure that the value received from this table is not -1,
because we don't want to try and access anything with the super class if we
can't distinguish which super class to use (because -1 doesn't correspond with
anything).