Allow my test case to pass!
With this change, I implement PRIM_SUPER_CALL in codegenPrimitive().  The
implementation is not neat - in fact, it implies that I should insert a
PRIM_GETCID call and then utilize that value as an argument to PRIM_SUPER_CALL.
But with this change, the test case that Engin sent me will now work as
expected.  Now to the clean up steps!