Modification to current solution to allow access of grandparent methods, too
Note - though this solution works when accessing grandparent methods, the
approach in this branch as a whole is flawed and leads to an infinite loop when
accessing an interim method.  I think part of this modification will be useful
in the proper solution, which is to determine statically which method to call.

IMPLEMENTATION:
Instead of using a binary "is there a super call involved" method of flagging
the access, count the number of supers involved in this access and utilize that
count to insert the appropriate amount of PRIM_SUPER_CALLs to get the right
cid.