Add "performance files" to enable better reproduction of bl.md failure
Verifies the behavior of the test in greater detail.  Turns on verbosity and
grabs the output for each of the verification steps encountered in the program.
Uses the performance validation for the final line, which ensures that if an
error occurs, we stop performing trials (and so the .exec.out.tmp is retained
instead of overwritten).