Insert default enum functions at scope where enum was defined
Previously, some of these functions were inserted in ChapelBase.
However, this meant that they would not get copied when the scope
in which the enum was defined was copied, leading to errors that
were worked around by explicitly inserting new versions.  This
change is the more principled approach, inserting the functions
at the same scope as the type itself.  With it, the workaround
can be removed.

Double checked over some enum tests, but still requires a full
paratest