Snapshot of first attempt, placed in driver.cpp
This seems like the wrong approach due to the need to tune what we will run the
subprocesses on - in order to make the executable/library pair, we need to know
the exported functions and that won't be possible until after parse has
finished.  But I think I will at least want the flags present, and it may be I
rethink this later.