Make --library-client and --library-server not infinite loop compilation
The pass to separate the library into a library/executable pair should not run
if we've already performed the separation.  Update --library-client to not
trigger the pass, and --library-server to generate an executable instead of a
library, even if --library is also passed (in case someone uses it by mistake or
I continue to be lazy and not remove --library from the arguments that get
passed to the server compilation).

With that done, actually run the subprocess compilation commands, and have the
separator pass exit compilation completely once the subprocesses have run.

Current status: failing to link C program with generated library file due to
undefined symbols