Snapshot, kinda works for using server function call template file
Only checked when function has no return type.  Was able to get replacement to
work (except arg handling) but only with the template text on one line.  Looking
into getting it to split onto multiple lines for readability, but that involves
more wrestling with sed to get it to do what I want.