Make printModuleInitOrder a compilation flag rather than an execution flag
When #14270 gets merged, it will add use statements to PrintModuleInitOrder by
necessity, which will modify the order of what gets printed and cause the
modules that it uses to not be printed at all (since their module initialization
functions are evaluated before its module initialization function, which sets
the variable that determines if the names are printed).  Making the config a
param instead allows them to still get printed out, but the presence of this
module still modifies initialization order.

As a result of this change, replace the .execopts for the test that tracks it
with a .compopts file.