Change the use of ChapelBase in ChapelArray to be private, with some fallout
Most places should have an explicit use of ChapelBase or get it from
ChapelStandard, so relying on ChapelArray to find its symbols seems like an
unnecessarily long use chain.

While here, I realized that ChapelArray was relying on ChapelBase to find
reductions, so added an explicit private use of ChapelReduce.

As a result of making ChapelArray's use of ChapelBase private, ChapelLocale
and DefaultRectangular needed to explicitly use it to access some symbols it
defines.  ChapelLocale also was relying on ChapelBase's use chain to access
memoryOrders, so add an explicit private use of MemConsistency to continue to
access it.