Use getVisibleFnsInstantiationPt instead of reimplementing the code it runs
Should result in the same behavior.  Don't do the same for
getVisibleFnsShowBlock due to that one explicitly putting visible fns in the
name when we say visible methods in the other branch.  This would be pretty easy
to unify, I just want to double check the instantiation point stuff works first