Initial implementation that works
Adds additional call path for forwarding operators as opposed to forwarding
regular methods.  Will go through every argument of the call and check
forwarding clauses for a call that will resolve.

Note: there is some duplicated traversal that occurs in the code as it stands
today.  I need to brainstorm how to fix that.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>