Fix bug where failed calls with forwarding were treated as cycles
We're only potentially in a cycle if there's still arguments left that can
forward.  By having the check outside the loop over the arguments, we were
accidentally marking failures to resolve as forwarding cycles, merely because
both the arguments were forwarded to the same type when we checked.  This change
moves the check inside the argument traversal and only when one of the arguments
can forward still - it also prevents us from now performing the same check for
every forwardable argument when the call itself won't have changed.

This allowed an additional test I've written to pass

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>