Make alternate cycle detection for forwarding operator calls
Operator calls will look for matches based on forwarding all real arguments -
when the operator is binary and both sides are of a forwarded type, this would
make it appear as though a cycle was occurring when really it was just because
we were checking the forwarded type for both arguments.  Now, treat it as a
cycle if we see more than two, but only for operator calls.

This resolved three tests I had written, which will be committed

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>