Add three tests of forwarding operators that appeared to be cycles (but weren't)
These tests involve various combinations of forwarded types that previously
triggered as cycles:
- One was a forwarded type that defines an operator on itself, and forwards to
another type.  This case triggered the old cycle detection because we'd see that
the type of the first argument could still be forwarded before checking if the
second argument was forwarded.  E.g., we'd check these cases in order:

Foo, Foo
 -> Bar, Foo
   -> Baz, Foo
   -> Baz, Bar
   -> Baz, Baz - and fail here, we'd think there was a cycle involving Bar
 -> Bar, Bar - if we continued on to this case, we'd find the match
 -> Bar, Baz - and the following cases would be unchecked
Foo, Bar
Foo, Baz

- The second and third tests were when the hierarchy of forwarding was similar,
but instead of defining the operator on `(Bar, Bar)`, we'd define it on a
combination of Bar and Baz.

With the cycle detection improvements, these all now pass.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>