Revert "Fix bug with returning the result of ternary expressions"
This reverts commit cba37d19a72e15edef4f7e25f4f0783b83a92112.

While this change managed to fix the bad behavior in some cases and left the
behavior in others unchanged, it did add memory leaks in some places that
weren't experiencing leaks otherwise.  Given that it was a quick fix to begin
with, it doesn't seem worth those new leaks.  We'll resolve this failure when
we re-implement the pass in dyno

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>